requirements.txt
pdfplumber
python-dateutil
pytesseract
pdf2image
Pillow
